YMD-GUI/requirements.txt
customtkinter==5.2.2
eyed3==0.9.7
pandas==2.2.3
Pillow==10.4.0
pytube==15.0.0
Requests==2.32.3
yt_dlp==2024.8.6
CTkListbox
